# gonna-pay-backend/docs/swagger.yaml
basePath: /
definitions:
  request.CreateContactRequest:
    properties:
      category:
        enum:
        - Family
        - Friend
        - Work
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    required:
    - category
    - email
    - name
    - phone
    type: object
...
  request.UpdateContactRequest:
    properties:
      category:
        enum:
        - Family
        - Friend
        - Work
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
...
  response.ContactFrequencyResponse:
    properties:
      contactId:
        type: string
      contactName:
        type: string
      sharedCosts:
        type: integer
    type: object
  response.ContactResponse:
    properties:
      category:
        type: string
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  response.ContactStatsResponse:
    properties:
      byCategory:
        additionalProperties:
          format: int64
          type: integer
        type: object
    type: object
  response.ContactsListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.ContactResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      stats:
        $ref: '#/definitions/response.ContactStatsResponse'
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  rest_errors.Causes:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  rest_errors.RestErrors:
    properties:
      causes:
        items:
          $ref: '#/definitions/rest_errors.Causes'
        type: array
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
host: localhost:3333
info:
  contact: {}
  description: API de gestão financeira e divisão de custos
  title: Gonna Pay API
  version: "1.0"
paths:
...
  /contact:
    post:
      consumes:
      - application/json
      description: Cria um novo contato vinculado ao usuário autenticado
      parameters:
      - description: Dados do contato
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.CreateContactRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ContactResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
      security:
      - CookieAuth: []
      summary: Criar contato
      tags:
      - contacts
  /contact/{id}:
    delete:
      description: Remove um contato. Apenas o dono pode deletar
      parameters:
      - description: ID do contato
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
      security:
      - CookieAuth: []
      summary: Deletar contato
      tags:
      - contacts
    get:
      description: Retorna um contato pelo ID. Apenas o dono pode acessar
      parameters:
      - description: ID do contato
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ContactResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "403":
          description: Contato não pertence ao usuário
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
      security:
      - CookieAuth: []
      summary: Buscar contato por ID
      tags:
      - contacts
    patch:
      consumes:
      - application/json
      description: Atualiza os dados de um contato. Apenas o dono pode editar. Campos
        são opcionais (patch)
      parameters:
      - description: ID do contato
        in: path
        name: id
        required: true
        type: string
      - description: Dados para atualização
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateContactRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ContactResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
      security:
      - CookieAuth: []
      summary: Atualizar contato
      tags:
      - contacts
  /contacts:
    get:
      description: Retorna todos os contatos do usuário autenticado
      parameters:
      - description: 'Número da página (padrão: 1)'
        in: query
        name: page
        type: integer
      - description: 'Itens por página (padrão: 20, máximo: 100)'
        in: query
        name: limit
        type: integer
      - description: Filtrar por categoria
        enum:
        - Family
        - Friend
        - Work
        in: query
        name: category
        type: string
      - description: Buscar por nome
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ContactsListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
      security:
      - CookieAuth: []
      summary: Listar contatos
      tags:
      - contacts
  /contacts/frequency:
    get:
      description: Retorna os contatos que mais aparecem nos custos do usuário autenticado
      parameters:
      - description: 'Número máximo de contatos (padrão: 5)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.ContactFrequencyResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
      security:
      - CookieAuth: []
      summary: Contatos por frequência
      tags:
      - contacts
...
schemes:
- http
securityDefinitions:
  CookieAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"